Add files via upload
I have made an e-commerce website using HTML, CSS, JavaScript, and Bootstrap 5. The website is designed for optimal responsiveness and a seamless user experience. It incorporates modern design principles, ensuring a visually appealing interface. Functionalities include product listings, a user-friendly shopping cart. The combination of HTML for structure, CSS for styling, JavaScript for dynamic elements, and Bootstrap for responsive design results in a well-rounded e-commerce platform. The website is a testament to the integration of front-end technologies to create a robust and visually pleasing online shopping experience.